Includes tvOS support for `UIAlertController`
