+++ Revert-tvOS/Resources/Data/AlertItems.plist
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<array>
+	<dict>
+		<key>rows</key>
+		<array>
+			<dict>
+				<key>title</key>
+				<string>UIAlertController: UIAlertView</string>
+				<key>cellIdentifier</key>
+				<string>alertcontroller</string>
+				<key>requiredClassName</key>
+				<string>UIAlertController</string>
+			</dict>
+			<dict>
+				<key>title</key>
+				<string>UIAlertController: UIActionSheet</string>
+				<key>cellIdentifier</key>
+				<string>actioncontroller</string>
+				<key>requiredClassName</key>
+				<string>UIAlertController</string>
+			</dict>
+		</array>
+	</dict>
+</array>
+</plist>
